+++ CHANGELOG.rst
@@ -14,6 +14,11 @@ and this project adheres to `Semantic Versioning`_.
 Unreleased_
 -----------
 
+Added
+.....
+
+* ``ensureAtLeastOne`` utility in evaluation context,
+    that will force property evaluation until a given predicate gets satisfied
 
 Dependencies Updated
 ....................